--- src/components/Testimonial/testimonial.css
@@ -1,11 +1,0 @@
-.carousal > img {
-    margin-top: 5rem;
-    width: 50px;
-    height: 50px;
-}
-
-.testimonial-carousals {
-    background-color: #041629;
-    color: white;
-}
-
