+++ src/pages/Testimonial/testimonial.css
@@ -1,0 +1,44 @@
+.carousal > img {
+  justify-content: center;
+  height: 50px;
+}
+
+.testimonial-carousals {
+  display: flex;
+  align-items: center;
+  height: 100%;
+  min-height: 70vh;
+  background: url("../../assets/png/banner.png");
+  justify-content: center;
+  background-size: cover;
+  box-shadow: inset 0 0 0 2000px rgba(0, 0, 0, 0.7);
+  color: white;
+}
+
+.carousal {
+  display: flex;
+  flex-direction: column;
+}
+
+.testimonial-text {
+  padding: 0 20%;
+  font-size: large;
+}
+
+.testimonial-text > h4 {
+  font-size: 1.5rem;
+  text-transform: uppercase;
+  font-weight: 600;
+}
+
+.testimonial-text > h6 {
+  font-size: 1rem;
+  font-weight: 600;
+  margin-top: -3%;
+}
+
+.carousal-image > img {
+  width: 50px;
+  height: 50px;
+  margin: 2% 0 -2% 0;
+}
